Pass Against Source Code
Previously, when this buildpack ran against certain kinds of source code, it
would error rather than always passing because multiple "scripts" would be
found.  Given that this wouldn't be what the compiled artifact would look
like, this failure was incorrect and prevented the buildpack from running
during the build phase when it could accurately determine if it should
participate.

This change updates detect to ignore any error that it might find so that the
buildpack can participate at build time against a potentially compiled
artifact.

Signed-off-by: Ben Hale <bhale@pivotal.io>