Docker Login Improvements
Previously the Docker Login step in many of the workflows was executed
regardless of whether it was necessary.  This would have been find except that
the required secrets were not available in PRs from forks.  This change
removes the Docker Login steps from where it is not necessary and puts a guard
on its execution in workflows that are run as part of PRs.

Signed-off-by: Ben Hale <bhale@vmware.com>