Require jvm-application-package in buildplan
Previously we used the jvm-application build plan entry to represent two concepts

1. package containing application bytecode
1. process types and configuration that will launch JVM application in the final image

This PR uses a new entry called jvm-application-package to represent application bytecode. Creating this distinction makes the role a given buildpack plays clearer and solves a practical problem. Previosly, when building from source code the jvm-application plan entry was being met by a build system buildpack (e.g. maven) and therefore not passed to buildpacks that contribute start commands (like dist-zip). Therefore, if a downstream buildpack wanted to add metadata to the jvm-application plan entry, intended to modify the behavior of a JVM application buildpack, that metadata would not be present if the image was built from source.

This depends on coordinated changes in the build system buildpacks.

Signed-off-by: Emily Casey <ecasey@vmware.com>