Add support for reloadable process types RFC
- If `BP_LIVE_RELOAD_ENABLED` is set and true
 - the CNB will require `watchexec`
 - the CNB will add a process type called `reload` that calls `watchexec -r <normal command>`
 - the CNB will set the default process type to be `reload`

This can be used in conjunction with a Tiltfile like:

```
custom_build(
   'example-java-image',
   'pack build --pull-policy=if-not-present -e BP_LIVE_RELOAD_ENABLED=true -e BP_GRADLE_BUILT_ARTIFACT=build/distributions/*-boot-*.zip',
   ['pom.xml', 'build/libs/demo-0.0.1-SNAPSHOT.jar'],
   live_update = [
       sync('build/libs/demo-0.0.1-SNAPSHOT.jar', '/workspace/demo-boot-0.0.1-SNAPSHOT/lib/demo-0.0.1-SNAPSHOT.jar'),
   ],
   tag="dev"
)
k8s_yaml('kubernetes.yaml')
k8s_resource('example-java', port_forwards=8000)
```

When code changes are made to a watched directory, Tilt will push the changes out to the container. Those will be observed by `watchexec` and the main process will reload, reloading your changes. With DistZip this is a little bit less practical because you are pushing up a ZIP file that contains a JAR file with all of your application files in that. You cannot easily recompile and upload a few class files. It is technically possible though to recompile and send your entire JAR file. When done, `watchexec` will see the change and reload the application.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>