Make script resolution more strict
Presently, script resolution looks at the default pattern of `*/bin/*`. This matches any zip with a bin directory and one or more files in it. This can incorrectly flag something like a Liberty packaged server as a dist-zip archive.

This change makes the script resolver more strict. It will only pass if there is a single script matched by the pattern. It seems that based on the [dist-zip](https://docs.gradle.org/current/userguide/application_plugin.html#sec:the_distribution) format should in most cases only generate a single Unix and a single Window script. The format is very flexible, so it's possible a user could customize and generate more scripts but this is going to be less common. Thus by default, we are only going to match when a single script is present.

If multiple scripts are matched, there will be messages emitted indicating what happened at the DEBUG log level. If the Paketo DistZip buildpack does not run (or does not emit it's title) at build time, look at the debug logs from detection to see if it matched multiple scripts. A user can then set `$BP_APPLICATION_SCRIPT` to a more strict pattern (or the exact file name of their dist-zip archive).

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>