Mark the detected script executable
When the buildpack runs, it will locate the dist zip script to execute. If that script is found, then we also mark it as `0755`. This happens unconditionally, but we do not fail if there is an error. We just log a warning and proceed. It really shouldn't fail but maybe if the file had weird file permissions/ownership.

The choice to do this unconditionally is that it is slightly simpler. Checking and then setting is two operations. If thereis a preference to doing it that way, making the switch is easy enough. Just let me know

Signed-off-by: Daniel Mikusa <dan@mikusa.com>